+++ Sources/KeyboardKit/Callouts/InputCalloutContext.swift
@@ -9,21 +9,21 @@
 import SwiftUI
 
 /**
- This context can be used to control input callout views, to
- present the currently typed character.
- 
- You can inherit this class and override any open properties
- and functions to customize the standard behavior.
+ This context can be used to handle input callouts that show
+ the currently typed character.
  
  `KeyboardKit` will automatically create an instance of this
- class and bind it to the input view controller.
+ class and bind it to the input view controller. The default
+ instance will only be enabled for iPhones.
  */
-open class InputCalloutContext: ObservableObject {
+public class InputCalloutContext: ObservableObject {
     
     
     // MARK: - Initialization
     
-    public init() {}
+    public init(isEnabled: Bool) {
+        self.isEnabled = isEnabled
+    }
     
     
     // MARK: - Properties
@@ -33,7 +33,7 @@ open class InputCalloutContext: ObservableObject {
     /**
      The optional input of any currently active action.
      */
-    public var input: String? { action?.input }
+    public var input: String? { action?.inputCalloutText }
     
     /**
      Whether or not the context is enabled and has an input.
@@ -41,10 +41,10 @@ open class InputCalloutContext: ObservableObject {
     public var isActive: Bool { input != nil && isEnabled }
     
     /**
-     Whether or not the context is enabled and will show any
-     callout. This is true by default for phones, else false.
+     Whether or not the context is enabled, which means that
+     it will show callouts as the user types.
      */
-    @Published public var isEnabled: Bool = UIDevice.current.userInterfaceIdiom == .phone
+    @Published public var isEnabled: Bool
     
     /**
      The action that is currently active for the context.
@@ -60,38 +60,19 @@ open class InputCalloutContext: ObservableObject {
     // MARK: - Functions
     
     /**
-     The visible button frame for the button view's geometry
-     proxy. You can apply an inset by subclassing this class
-     or adjusting the style.
-     */
-    open func buttonFrame(for geo: GeometryProxy) -> CGRect {
-        geo.frame(in: .named(Self.coordinateSpace))
-    }
-    
-    /**
-     Reset the context, which will reset all state and cause
-     any callouts to dismiss.
-     */
-    open func reset() {
-        action = nil
-        buttonFrame = .zero
-    }
-    
-    /**
      Update the current input for a certain keyboard action.
      */
-    open func updateInput(for action: KeyboardAction?, geo: GeometryProxy) {
+    public func register(_ action: KeyboardAction?, in geo: GeometryProxy) {
         self.action = action
-        self.buttonFrame = self.buttonFrame(for: geo)
+        self.buttonFrame = geo.frame(in: .named(Self.coordinateSpace))
     }
-}
-
-private extension KeyboardAction {
     
-    var input: String? {
-        switch self {
-        case .character(let char): return char
-        default: return nil
-        }
+    /**
+     Reset the context. This will cause any current callouts
+     to be dismissed.
+     */
+    public func reset() {
+        action = nil
+        buttonFrame = .zero
     }
 }